Broke log loading into independent script
